Object Mimicry Work
I tried seeing if separating the colliders from the graphics objects performed better for the mimicry, however it just made it more complicated and I ended up reverting it.

Now, the MimicableObject's graphics prefab should be the scene-instance version of their graphics, however we can change this later if it would be easier/more performant.