Mimicking Rigidbodies
I have made it so that when a Mimic mimics an object, they also copy that object's rigidbody information to their own (If the target object has a Rigidbody).

Furthermore, I have the mimic store its own rigidbody information in the Awake call, which it then uses when reverting to its original form.