Stalking Behaviour Start
I have started work on the Stalking Behaviour, making a system to allow us to create Hiding Spots which the agents can move to in an attempt to remain hidden.

Hiding Spots update to record whether they would be in view of the player if there were no obstructions (So that we don't have the issue of the player rounding a corner and seeing the creature standing with its back to them, thinking it would have been hidden), their square distance to the player, whether the spot is too far from the player, and whether the spot is occupied (In the event that we want multiple agents).
Hiding Spots are currently sorted by their square distance with the closest valid spot picked by agents when they are searching for a new one. Stale spots aren't marked as invalid, and instead notify the agent via a separate event so that they can wait until they aren't going to be observed when they leave.