AStar Pathfinding Start
I have started work on implementing a A* Pathfinding algorithm by following Sebastian Lague's tutorial on the subject.

I am doing this so that we have better control over the generation and manipulation of the entity's pathing, which we can use for stuff such as ensuring that they don't path in front of the player unless necessary.