Teleporter Rotation
I have made it so that teleporters maintain the player's relative rotation when teleporting (E.g. If you are facing 90 degrees to the right relative to the forwards of a teleporter, then you will exit facing 90 degrees to the right of the exit teleporter).

Furthermore, teleporters now work using an ITeleportableObject interface.