Teleporter Base Class
I have created a BaseTeleporter class which both types of teleporter inherit from. This allows us to reuse some of the code (E.g. For Warmup, Starting, and Stopping) and should make maintenance and changing easier in the future.

I have also prepared the Inter-Scene Teleporters a little bit more for the Scene Management.