Scene Management Start
I have started work on the Scene Management & Transitions, getting it to the point where we can load & unload scenes successfully.

We need to determine if we want to instead use a Singleton and Persistent Scene rather than a static class, as it may be easier for having a Loading Screen.