AStar View Avoidance Start
I have implemented a basic version of the A* Viewcone avoidance by having the pathfinder simply discount nodes which are viewable.

This does mean that if the only path to a target is in the player's view, no path will be found. If this isn't desired, we can implement node weighting and have visible nodes instead have a much higher cost.