EditorInitialiser Fix & Player Prefab
I have created a Player prefab based on the one Braidyn was using in their scene and have added it to the Persistent Scene. Due to the existence of the Persistent Scene, I had to rework some of the scripts which directly referenced the player, and have temporarily strung them around the PlayerManager to get the player's transform.
(A similar thing has been done with the Flashlight & Charging Station, however these are using a 'TmpFlashlightController' which holds a reference to the 'FlashlightHolder' transform.

Furthermore, I have fixed a bug with the EditorInitialiser where it would always load the PersistentScene, even if in a scene where we didn't want to load it. This was due to the fact that we were never resetting the EditorSceneManager's playModeStartScene from the PersistentScene when the active scene was invalid.