Interactables & New Input System
We have changed the Interactables system so that it uses an Interface instead of directly having logic for every different type of interactable. This also allowed us to easily integrate the FlashlightPickup and RechargeStations into the interactables system (So they no longer rely on Trigger Colliders & Tags).

Additionally, we have made it so that all our current scripts work using the New Input System with our PlayerInput script as a mediator. This required a slight reworking of the scripts, but I have kept as much of people's old code and comments as I could.

Furthermore, I made the following changes/bugfixes:
- The FlashlightPickup (Previously FlashlightPickupController) no longer requires the Flashlight Logic to function.
- The FlashlightRechargeStation (Previously FlashlightRechargeController), PlayerHiding, & PlayerInteraction scripts no longer rely on string-based tags (This is a benefit as it reduces the chance of errors when typing in tags, due to them being weakly typed)
- Adjusted the Player's NormalHeight and CrouchHeight values in their PlayerController script.
- Added in the ability to separately configure your horizontal and vertical look sensitivity.
- The player's look input is now scaled with Time.deltaTime, making it framerate independent.
- Added in the option to use Toggle Crouch and Toggle Sprint (Toggle Sprint automatically ends if you stop moving)