ItemSpawnManager & Hiding Spots in Level
I have created an ItemSpawnManager script to facilitate the random spawning of keycards, along with the overriding of materials for Interactable Doors and Keycard Readers while testing to match the colour of their associated Keycard (There are also checks to notify us if we add a Keycard with no associated Door/Keycard Reader and vice versa). I also removed the material colour changing from the Keycard class, as it no longer functioned within the URP and was not necessary with the ItemSpawnManager overriding the material anyway.

Furthermore, I have added in some of the Table Hiding Spots into the Stealth Demo scene.