Navigation & Senses Start
I have started work on the Mimicry Controller rework and the Mimic's senses (Player Detection via Sight & Sound). Furthermore, I have installed the AI Navigation package along with creating a Navigation Surface for the level with the doors as obstacles so that they can be toggled to let the Mimic through when unlocked/open.