Player Controller Behaviour Fix & Crawling
I have fixed issues with the Player Controller which resulted from being able to change states when you weren't supposed to (E.g. Stopping crouching when there is something above you). Furthermore, we added a transition period between height changes for the CharacterController & Camera so that it isn't as jarring.

Additionally, we have implemented crawling & the vents, triggered when the player is crouching nearby a vent.