Flashlight Recharge Station Change
We have changed the Flashlight Recharge station to use a fixed rate rather than a fixed time, meaning that it takes less time to charge a flashlight to full if it is closer to full to begin with.

Additionally, we have made it so that you can pick up the flashlight while it is charging, interrupting the process but letting you use it early.