Saving to & Loading from files start
We have made a start on Saving to/Loading from .json files, with separate naming conventions for Autosaves and Manual saves, the ability to select the most recently created file, and within that the most recently saved element (Manual, Checkpoint, Hub, etc).

Currently these are called via the inspector and aren't hooked up to the actual saving, and there seems to be an issue where the flashlight battery doesn't update.