Saving Improvements & Passive Mimicry Error Fix
We have improved the saving system so that we can autosave after a constant delay (For testing it's currently every 0.5 minutes (30s), but we can increase it to every 5mins or something once testing is completed).

Additionally, we now save the data from when a save occurs and use that when reloading.

Furthermore, we have fixed a bug with the Passive Mimicry Camera which caused errors to throw when the scene is reloaded.