Main Menu & Load Save Menu
We have started work on the Main Menu scene, which will have functionality for things such as Starting New Games, Loading Existing Games (Most recent or specific save), Accessing the Settings Menu, and Exiting to Desktop.

We currently load the PersistentScene from the MainMenuUI script (Which is alright), and the New Game/Continue/Load functionality hasn't been implemented yet, and we have a temporary bodge for stopping the SaveManager from performing autosaves in the Main Menu (Which we should improve).