Start Buttons & Bugfixes
We have implemented the functionality for the 'New Game', 'Continue', and 'Load Game' buttons.

Additionally, we have fixed several bugs, including the player's flashlight battery not getting set to what it is in the save data and the player spawning standing in vents when the save occurred inside of one.
Furthermore, we allowed ourselves to sort our obtained save files in descending order of their creation time (Index 0 is the newest file)