New Action Maps & Action Prevention
We have separated the player's action maps into 4 types: Movement, Camera, Interaction, and Inventory.

We have also made it so that we can enable & disable these maps depending on the UI elements which are active, allowing us to (For example) prevent all input when the Loading Screen is up, or prevent camera and interaction input while the Inventory is open.