Input Prevention Fix
We have fixed a bug where an error would throw if the PlayerInput class hadn't yet generated when a UI element wished to disable an input map. We now check this before enabling/disabling, and when the PlayerInputActions map is generated we check and disable elements that have preventions associated with them.

Furthermore, we removed the singleton aspect of the PlayerInput class, as we no longer require it.