Bugfixes
Removed the autosave on first load - We need to be a bit more careful when testing directly from a scene, but this was throwing errors on a regular load.

Mimic now kills only once - The mimic was previously killing the player over and over again, causing the GameOverUI to sometimes get stuck active when reloading the scene.