Main Menu Don't Destroy on Load Removal
We've changed the SettingsManager and SettingsUI so that they don't call DontDestroyOnLoad() for the object they're sitting on.

We have also changed the PauseMenuUI so that it isn't just a paste of the Main Menu's hierarchy, but instead is the pause menu. Additionally, we made it support the NewInputSystem, got it started on Gamepad Support, and fixed issues where the pause menu would be non-functional due to DontDestroyOnLoad.