Player Settings Change
We have changed how the Player/Controls settings work in the SettingsUI.
Now, we have a static PlayerSettings class which contains values for the player's settings, such as Mouse Sensitivity, Gamepad Sensitivity, Toggle Crouch, etc. Additionally, this class handles the loading & saving of data from PlayerPrefs.

Furthermore, we can now set the toggle state of the inventory (Hold to open or press to toggle open state), along with configure separate values for the sensitivities of the mouse & gamepad input, along with separate values for their vertical & horizontal axies.