Demo Security Level Fixes (& Empty Lightmap)
We have fixed the Security Levels of the Keycard Readers within the three demos, along with remaking the Chase Demo so that the player has slightly more chances to make mistakes before they die.

Additionally, we have created an empty scene and baked a very low size lightmap from it, which we can use in scenes to give an ambient environmental light without needing to actually bake the scene.