Script Re-organisation Completion & Hiding Spot Change
We've finished the re-organisation of the project's scripts, however we have still yet to go through the scenes to remove missing script references & re-assign values.

Additionally, we've changed how the Hiding Spots work, meaning that the PlayerInteraction script no longer needs to know about HidingSpots or the PlayerHide script. Furthermore, we can now specify the character height and camera height to use when hiding, where we can exit to, and we can use curves for the transition into and out of a hiding spot.