Scene Fixes & Hiding Spot Prefab
We have went through and fixed (Or deleted) all the scenes with errors caused due to the changing of scripts, along with completing a full playthrough of each demo scene accessed through the Prototype Hub.

Additionally, we have added the option for Hiding Spots to automatically delete obstructed exit positions on game start, letting us use a prefab for all of them. Subsequently, we have created a prefab for the table hiding spot, and replaced the old tables with it.