Merge Fixes, LevelLayout Organisation, & Mimic Prefabs
We have fixed the errors caused when merging the ScriptCleanup branch into main.

Additionally, we have also organised the hierarchy of the LevelLayout, along with merging the walls, ground tiles, and ceiling tiles together to clean up the inspector. (Note: We haven't yet removed the harsh directional light, as doing so made viewing the scene near-impossible)

Furthermore, we have created 2 prefabs for the mimic: One for Chasing and one for General. These prefabs both have the new mimic model as their GFX.