Reloading Fixes & Exit To Main Menu Implementation
We have fixed bugs with the Reload to Checkpoint and Reload to Hub reloading options which were causing issues and would occasionally cause the game to crash.

We have also implemented the 'Exit To Main Menu' feature, letting the player actually quit to the main menu instead of the button doing nothing, and have added the option to exit to the main menu from the Game Over screen.