Keycard Decoder Level Load Reset Option
We have implemented the ability to have the KeycardDecoder reset its security level to 0 when a new scene is loaded. This will be useful for the playtest so that the demos remain distinct and don't conflict with each other, but as it's togglable we can remove it when we get to testing our combined levels.