PSX Shader
We've re-implemented the PSX Shader that Lucy found and fiddled around with its settings for a bit, getting it to a point that is in an alright place.

We've also had to remove the Overlay Camera from the player (As the PSX effects were applying twice), so the player's items have been scaled down so that they fit inside the player's collider, and therefore won't clip.