Hallway Scene & Shader Prep for Build
We have created a Hallway scene and allowed the user to control when the Mimic walks past (A, D, & S), and whether they are cloaked or not (Q & E).

Additionally, we learned that you need to move the Shaders folder into the Resources folder in order for the Post Processing ones to be included.