LevelLayout Fixes, Start Transition Position Change, & Model Implementations
We've fixed some issues with the Level Layout scene, including the missing GeneralMimic, gaps above the doors, and the mimic being able to kill you in the teleporter.

We've also fixed the LevelLayout_StartGame transition so that the player spawns in the starting room rather than in the void.

Furthermore, we've implemented Lucy's models and materials for the Keycard Pedestals, Keycard Readers, Keycard Pickups, and Pipes, however models such as the Teleporter had issues with their normals.