UltEvents
We've imported the UltEvents asset from the Unity Asset Store, which allows us to have more robust and feature-complete, while still Designer-Friendly, versions of UnityEvents. One of these features (The main reason we're using them) is the calling of Static methods, which will be utilised for stopping chases without requiring a direct reference to the chaser).