Interaction Trigger & Volume Trigger
We have implemented two types of ScriptedEventTriggers, both of which use UltEvents in order to allow for easy designer implementation, modularity/reusability, and the calling of static events.

We have implemented a 'InteractionTrigger' which is placed on an interactable (E.g. A KeycardPickup) and calls its linked events when an interaction is performed (With options to only trigger when the interaction was successful or a failure, or either)

We've also added a Volume Trigger, which calls its linked events when one of a set type of collider enters its trigger volume. The current types are: Player (Contains a PlayerController component), Mimic (Contains a GeneralMimic or ChaseMimic component), and Object (Contains a non-kinematic rigidbody)