Preset Additions, ChaseDemo Door Fix, & Door Initial State Set Fix
We've added the option to use add a preset event call to the ScriptedEventTriggers by right clicking on the name of the class in the inspector. The current presets are for triggering a chase end immediately and triggering a chase end after a delay.

Additionally, we fixed the Chase Demo's doors so that they now work with the scripted event triggers, removed the old scripts, and fixed a bug in the Door class's initialisation where some doors would initialise before their DoorOpener scripts would subscribe to them, leading to them setting their initial state before the DoorOpener class was ready.