VisibilityTrigger
We've implemented a VisibilityTrigger: A ScriptedEventTrigger which uses the player's camera to determine if it should activate.

We can set a maximum activation distance, whether it should only activate if unobstructed, and the horizontal and vertical screen scale within which it activates.

An example use case for this is a creature leaning around a corner, visible in the corner of the player's eye, before scattering away when the player looks at them.