Pipe Jumpscare change & Only Destroy Script option
We've changed the PipeJumpscare script to not handle its own activation, instead utilising a ScriptedEventTrigger to do so (Such as the VolumeTrigger, which replicates the previous implementation).

Additionally, we've added the option to only destroy the ScriptedEventTrigger script for when it is on objects that are not to be destroyed, such as the PipeJumpscare.