Venting Tests
We've started work on implementing the tests for the Mimic being able to move freely through vents, currently using a tester script to do so.

We're going to want to test out how to manually control the position of the mimic when it is entering a vent so that we can specify what position change it should make for handling animations.