Manual Link Movement & Gradual Height Scaling
We've implemented the ability to manually control the lerp-rate of our agent's horizontal position when they are entering/exiting a vent, including the option to have different curves for the different directions.

Additionally, we've added the temporary option to specify a curve for the height change, however for the proper implementation this would instead be handled by our given animation (We may want to use this for colliders or similar though).