Sight Retention, Chase Exit Fix, & Rotation Fix Failure
We've implemented sight retention, allowing the mimic to know the player's position for a short duration after losing sight of them.
Additionally, we added the option to ignore the view angle when determining if we've lost the player, preventing the mimic from losing sight of the player if they dip behind them.

Furthermore, we have fixed the transition from the Chase state which was causing it to exit as soon as they lost sight of the player, as opposed to when they reach the last visible position.

Finally, we attempted to fix the mimic's rotation so that it would stop running into objects, however we failed to do so and will need to look into this more.