Crawlable Layer Cost Change & Mimic Agent Radius Increase
We've decreased the cost of the 'Crawlable' navigation layer from 3 to 2, which is closer in cost to the 1/2 speed penalty applied to the mimic when crawling.

Additionally, we've also increased the radius of the mimic's navigation radius from 0.2f to 0.3f, which should hopefully reduce the possibility of clipping when moving near corners.