Footstep Clips (Carpet + Metal Walking)
We've added in footstep clips for Walking & Running on 'Carpet' as the default footstep sound, along with clips for walking on metal which is currently being used for Crawling.

We will later introduce footstep sound separation separated by Material Type as well, but for now they're only separated by MovementState.