Animation Events, Mimic Footsteps, & SFXManager Rework
We've implemented a different kind of Animation Events, allowing us to trigger events using string identifiers rather than requiring us to have an instance of a component on the animated object.

Furthermore, we've added in temporary animations for the mimic which should be replaced with the actual versions as soon as we can, and through these and the Animation Events we have implemented the mimic's footsteps.

Furthermore, we have changed the SFX Manager to use a pool of AudioSources rather than a single AudioSource with PlayOneShot, allowing us to have customisation on each without effecting the active clips.
Furthermore, we can now also influence the falloff curve of the AudioSources.

Temporary Scripts still in project:
- AnimationAudioTrigger.cs