Equipping & Canvas Post Processing Fix
We've implemented the ability to equip the Player's Tablet through a button press (Escape, same as pausing), along with adding a second camera to the player which only renders the UI, allowing us to disable Post-Processing for that camera, making the tablet's information legible.

Note: We've disabled the PauseMenu's subscription to input, meaning that you currently cannot activate it.