Player Tablet Settings Menu & Quitting, + FPSDisplay Bugfix
We've implemented the Settings Menu into the Player's Tablet, along with adding in Quitting & Checkpoint Reloading (Though not Save Reloading) with confirmations.

Additionally, we fixed a bug with the FPS Display where it wouldn't update its active state when the game was first loaded, and we've made it so that the SettingsMenuUI doesn't require a reference to it in order for the toggling to work.