Collectable Collection & SerializableGUIDs
We've implemented a CollectableManager to hold the current collected collectables, currently sorted by type and ordered by an integer value.

We can reference this data to collect a list of any desired type (So long as said type inherits from CollectableData), and are using this to display the currently collected data in the Player's Journal.

Furthermore, we've implemented 'SerializableGUIDs' from one of 'git-amend's tutorials, which we were originally going to use for saving the Collectables, however this currently isn't implemented.