Collectable Ordering Improvement
We've improved the way that we determine the sorting order for collectables, now having a ScriptableObject that contains a list of the Collectables, using their position in the list to determine sorting order.

This means that we don't need to remember what integer values have already been set, and there are errors and warnings to prevent us from making a mistake when setting things, and fallbacks so that those mistakes aren't fatal.