Game Over Screen Bugfix & Player Rotation Loading Fix
We've fixed a bug with the PlayerHealth script and the GameOverUI which was causing the GameOverUI to remain up after reloading (The player was dying every frame, including the frame between the UI going down initially and the game being reloaded).

We've also fixed a bug with loading the player's rotation where we were saving and loading the player's root's rotation (Which was always 0,0,0), as opposed to the player's Rotation Pivot's rotation (Which is what actually changes).
We also updated the ForegroundSceneTransitionSO to match this, so the rotation setting for that should work now too.