Settings Menu UI Rework Completion
We've finished reworking the Settings Menu UI for the Player Tablet & the Main Menu, both of which now work and are navigable via controller.

Furthermore, we've adjusted the scaling a bit more so that the Player Tablet is more legible, even with a small screen size.

Note: When the player's FOV changes, the tablet breaks.