Shader Toggling & CRT Feature Removal
We've implemented the ability for us to toggle the PSX Post-Processing Features by making it so that they are only active if there is a Post-Processing Volume in the scene which has overrides applied to the desired feature.
This allows us to entirely disable them without needing to go into the URP settings, however it also means that they only display if there is a Post-Processing Volume in the scene.

Additionally, we've removed some of the features from the CRT Effect, namely Vignette and Noise, as we already have those features in other Post-Processing Effects and by removing them we can make the effect of the CRT more clear.