ChaseMimic Activation Change + Pausing
We've changed how Chase Mimics are activated so that you now require a reference to the specific Chase Mimic you wish to activate. There are still static methods for Pausing, Resuming, and Ending all active chases.

Furthermore, we've added the ability to Pause and Resume Chases, with the option of keeping the chase music going while the chase is in progress.
(We also set the chase music in the Chase Mimic prefab)

Additionally, we made it so that you can specify if you'd like to disable certain actions during a Camera Focus Look, with the default being that we prevent the player's Movement during the duration.