CameraFocusLook Movement Direction Bugfix & Fighting Cancel
We've fixed a bug where the player's movement direction wasn't changing when their camera direction changed from the CameraFocusLook script.

Additionally, we've made it so that if we don't specify that we want to prevent camera input we initially disable it for the first X seconds, and then cancel the focus look if we have input after that time.