Door NavMesh Fix, Vents (Chase) Work, & Mimic Bugfixes
We've fixed the issue with NavMesh generation when Doors were present by adding a NavMeshModifier component to them set to remove.

We've also done some work on the Vents (Chase) scene, fixing issues with mimics pausing where they should end, giving the player some extra space at the start so they aren't instantly damaged, and adding in an initial door so that they cannot enter the vents without the required keycard.

Furthermore, we've fixed a bug in the MimicAttack script, along with removing its dependency on either the ChaseState or ChaseMimic class, so the mimic now properly attacks the player. We also made it so that we can specify when we are able to attack (Default is true), and made it so the GeneralMimic can only attack while in the ChaseState.

Additionally, we fixed several bugs relating to the GeneralMimic, including it not being able to see through doors (Even when they were open) due to Trigger Colliders, it using the wrong values when overriding acceleration (It was using the speed override), and implementing the ability to override the agent's stopping distance.
We then tweaked a few values to get it feeling better.