Lighting Distance + Angle Attenuation Change
We've changed the Lighting Distance & Angle Attenuation calculations in URP as per the following example: 'https://docs.unity3d.com/6000.1/Documentation/Manual/urp/lighting/custom-lighting-change-light-falloff.html'.

This makes the light falloff closer to what it was in the Built-in Render Pipeline and reduces the number of overexposed elements while still allowing us access to URP features.