Interaction Popup Work
We've created a "PopupTrigger" and "PopupSetupInformation" class to handle the inspector setting of the popup information. The PopupTrigger is still called through events from ScriptedEventTriggers, but this way makes it more legible when editing in the inspector.

We've also introduced the option to fade rather than disable when the player exceeds the max distance/is obstructed from the popup, allowing the popup to remain until the desired action is taken.