Dynamic Text Box Sizing & Keep Icon Centred Option
We've implemented dynamic sizing for the Popup Element's text boxes, which should hopefully prevent them from clipping into the icon when displaying their text.
Additionally, we've added an option for keeping the Icon centred, rather than having the centre of the object be based on its area, for the cases where we wish that to be the case.