Screen-Space Popups, Custom Text, Text Sizing, & Popup Fading
We've implemented Screen-Space popups, along with adding the ability to display custom text and sprites rather than the Icon associated with the desired action. We've also added in the option for setting the text's font size & sprite size (Width = Height), along with having the popups fade in and out when they are created and before they are released.