Vents (Chase) Scene Walls and Roofs
We've added walls and roofs to the Vents (Chase) scene, along with adjusting the position of the lights so they aren't clipping with the walls.
Furthermore, we've adjusted the final trigger so that the player has a better view of the Keycard Reader while still being able to see the Mimic.

We've also changed the ChaseMimic script temporarily to have the SFXManager handle playing the footstep clips and the break door effects, while (For now) keeping the background music on the ChaseMimic (Will be changed shortly, hopefully).