Background Music Manager Start & Singleton Base
We've started work on a BackgroundMusicManager script which will manage the background music and its overrides (E.g. Chase Music). Currently, we can fade between various clips, override the active clips and revert said overrides, and control the transition duration between clips.

Additionally, we've created an abstract base class for Singleton scripts so that we don't need to keep duplicating logic, however we've so far only used this in the BackgroundMusicManager and SFXManager scripts.