Save System Rework Start
We've started work on the rework of the Save System, making it more expandable and robust.

We currently only have player position, rotation, and movement state saving, but it should now be easier to add more things that we are wishing to save.

Additionally, we've reworked the menus so they already function with the reworked save system.