General Mimic Wander Areas
We've made it so that the General Mimic has areas which it can wander between, with the ability to check whether it can reach a specified point within said area in order to determine whether it should query it for wander checks (This state is updated every 5.0 seconds to reduce computations).

This should hopefully properly fix the issues we've been having, along with introducing a more versatile method for declaring where the mimic can wander.