Level Saving Start
We've started work on implementing Level Saving, leaving Progress Saving where it is for now.

We've made a LevelDataManager which will sit at the root of any scene we are wishing to save level data to, and while in the inspector we can press a button in the script's custom inspector to find all ISaveableObject component instances in the scene.
These are then used with some basic binding to get our save data, along with loading the save data through similar methods.
We can currently save values, along with whether this object was destroyed or not.

We haven't yet integrated this with the rest of the Saving System yet, but the skeleton is there.