Object Save Data Improvement
We've changed how the ObjectSaveData and ObjectSaveInformation containers work.

Now the ObjectSaveData container has arrays for bool values and int values, rather than referencing a ObjectSaveInformation, and ObjectSaveInformation is instead used to pass values to the ObjectSaveData, meaning that we no longer need to rely on constants and conversions in the inheritors of ISaveableObject.

This also fixes a bug with the previous commit where we couldn't load any information as the conversion to the required save data would fail.