LevelDataManager SaveableObjects Loss Bugfix
We've fixed a bug that was causing us to lose the LevelDataManager's SaveableObjects array when we changed/reloaded scenes due to it not registering the changing of the value via editor functions as an actual change.
We've fixed this by marking the object as dirty after the 'LinkSaveableObjects' function is called in the editor script.