Better Loading Bugfix & Keycard Reader Load Fix
We've implemented a better way for handling the issue of null reference exception during the loading of data by having the data load before the script initialisation delay, meaning that there is no time for the null reference exceptions to be triggered.

Additionally, we've fixed a bug with the Keycard Reader which was causing it to not properly load its unlocked state.