Scene Index Bugfix & Prefab Instance ID Fix
We've fixed a bug that was causing our scene index to reset to -1 when we entered play mode, causing our saving and loading to not function properly.

Additionally, we've changed how we're using IDs to fix the issue with prefab instances all sharing the same ID. Now, we use the Instance ID of the GameObject, along with additional random integers to ensure that no components overlap, in order to determine our IDs.