Serializable ID Fix
We've fixed the Saveable Object IDs by reverting to the SerializableGUID as the Instance ID of a GameObject resets near constantly (On scene reload, on editor reload, etc).

We are initialising the SerializeableGuid when you create an object, and we have also fixed the 'Regenerate All IDs' method in the LevelDataManager and moved it into the Editor for ease (With 2 levels of warnings to ensure that we don't accidentally perform the operation and break all existing saves).