Hiding Spot Falling Through Floor Fix
We've fixed a bug with the Hiding Spots where the player would occasionally fall through the ground.

This seems to have been a mix of the player's vertical velocity not being reset while hiding and the hiding spot putting the player's collider slightly into the ground as it didn't account for their skin width.