Game Over Screen & Quit Confirmation Fix
We've fixed some issues with the Game Over Screen, including it not showing when the player dies and it not disappearing after an option was selected.

Additionally, we've fixed the rest of the window still displaying when the Quit Confirmation was being displayed in the Player Tablet. It now only shows the Quit Confirmation, and we've adjusted the text to make it display a bit nicer.