Lock & Map Clipping Fix & Improvements
We've fixed the issue of the Padlocks and Map clipping into the environment when interacted with at specific angles.

Additionally, we've improved the scripts a little, allowing the player to move the map when they have it selected, and having both items use the same 'FocusableObject' script to handle their focusing to reduce code re-use (Letting us add some niceness), and we've fixed some issues with how we were calculating the target position & rotation.