Save Loading Fix, Activate Trigger On Save Load Removal, & Name Changes
We've fixed a bug with the Save Loading which was causing us to re-initialise our 'BoolValues' and 'IntValues' arrays when we were loading saved data, causing us to lose that data and lots of errors to be thrown. Save Loading now loads data correctly.

Additionally, we've removed the '_triggerEventIfDestroyedOnLoad' option from ScriptedEventTriggers, as we aren't using it and it was causing issues of triggers playing on game load unexpectedly.

We've also changed how we're naming the save files, now in the form 'Playthrough-X_Autosave'/'Playthrough-Manual', where 'X' is the SaveID: a positive integer.

There is a known issue of Visibility Triggers triggering when we load.