Keycard Decoder Screen UV Fix & Various Save Fixes
We've fixed the UVs for the Keycard Decoder's screen, now letting the shader display properly.

We've also done various fixes in the Save System, preventing bugs where the Level Save Data wasn't reverting on a new game start, and InteractionTriggers weren't properly setting their Destroyed state due to OnDisable being overriden.

Furthermore, we've fixed a bug with the Player Interaction which was causing an error when trying to reset the current interactable override if we didn't already have one.