Repair Spot Bugfix & UI Colour Implementation
We've fixed a bug with the Repair Spots by changing the win detection to be based off of their completion progress rather than an integer counter (As it is more robust and the performance impact is negligible).
The bug that was fixed was being unable to add a Key Item to the escape pod, and in our testing it has no longer appeared.

Additionally, we've implemented the setting of the KeyItemUI colours for if it is holding the currently required key item or not (Rather than always showing red).