Intro Cutscene No Light Fix
We've fixed the Intro Cutscene not having any lights when you load into it from another scene.

It was caused by the Level Partitioning, and we've resolved it by having the Intro Cutscene add a count to the hub's perceived player count while it is active, removing it when it stops.