Passive Mimicry Change & Table Transparency Fix
We've changed the Passive Mimicry shader so that it no longer requires a second camera to render, instead using the 'SceneColor' node. This dramatically improves the performance, and should also fix the bug of the Passive Mimicry displaying the skybox sometimes.

We've also fixed the Hiding Spot Table's materials being transparent, which was causing the new body materials to draw in front of it.